Nivel 10 fuera del panel: el mixed que nacia dentro, a cero
El nivel 10 cierra el mixed que el propio codigo fabrica: closures sin tipo, static de funcion,
getters llamados por nombre, (array) de un objeto. Eran 111 fuera de los CRUD de EasyAdmin.

- TarifaDailyPriceFlattener lee el rango una vez a una forma con tipo; 3000 combinaciones
  aleatorias dan el mismo ganador y el mismo sourceId que antes.
- MessageTemplate: los setters del CollectionField guardaban (array) sin mirar, y el tipo no
  declaraba origenHash, la huella de la autotraduccion que traen las filas reales. Declarada y
  conservada; las tres entradas de baseline que protegen esos guardas cambian solo su texto.
- normalizarParaDB() exige lo que promete (las 3032 filas guardadas pasan).
- transliterate() devolviendo false dejaba vacio el nombre de un archivo del paquete de escaneos.
- Un expiresIn no numerico de Beds24 daba una caducidad basura al token.

Quedan 267, todos en controladores CRUD del panel.

+++ src/Agent/Skill/Pms/ConsultarDisponibilidadSkill.php
@@ -481,14 +481,14 @@ final readonly class ConsultarDisponibilidadSkill implements SkillInterface, Ski
      * libres, y qué preguntar para poder cerrar. Sin la última se queda en un dato y lo que
      * hay que hacer es vender.
      *
-     * @param list<object> $libres
+     * @param list<PmsUnidadDisponibleDto> $libres
      */
     private function reparto(array $libres, ?int $pax): string
     {
         $capacidad = 0;
 
         foreach ($libres as $u) {
-            $capacidad += (int) ($u->capacidad ?? 0);
+            $capacidad += $u->capacidad ?? 0;
         }
 
         $alcanza = $pax !== null && $capacidad >= $pax;
